p7microgrid/microgrid-starter-denmark/requirements.txt
python-microgrid==1.4.1
pandas>=2.0
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
PyYAML>=6.0
